requirements.txt
numpy==1.23.4
pandas==1.5.1
python_Levenshtein==0.20.9
torch==1.9.0
tqdm==4.64.1
